+++ src/modules/error/error.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
 
 /* Modules */
 import { AppCommonModule } from '@common/app-common.module';
@@ -27,7 +26,6 @@ import * as errorServices from './services';
         RouterModule,
         ReactiveFormsModule,
         FormsModule,
-        HttpClientModule,
         AppCommonModule,
         NavigationModule,
     ],