+++ src/modules/app-common/services/app-common.service.spec.ts
@@ -9,7 +9,7 @@ describe('AppCommonService', () => {
         TestBed.configureTestingModule({
             providers: [AppCommonService],
         });
-        appCommonService = TestBed.get(AppCommonService);
+        appCommonService = TestBed.inject(AppCommonService);
     });
 
     describe('getAppCommon$', () => {