+++ src/modules/navigation/components/breadcrumbs/breadcrumbs.component.spec.ts
@@ -48,7 +48,7 @@ describe('BreadcrumbsComponent', () => {
         component = componentDE.componentInstance;
         componentNE = componentDE.nativeElement;
 
-        navigationService = TestBed.get(NavigationService);
+        navigationService = TestBed.inject(NavigationService);
 
         fixture.detectChanges();
     });