+++ src/modules/navigation/containers/top-nav/top-nav.component.spec.ts
@@ -45,7 +45,7 @@ describe('TopNavComponent', () => {
         component = componentDE.componentInstance;
         componentNE = componentDE.nativeElement;
 
-        navigationService = TestBed.get(NavigationService);
+        navigationService = TestBed.inject(NavigationService);
 
         fixture.detectChanges();
     });