+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Changed] Only import HttpClientModule in appModule
+
 ## [1.1.0] - 2020-02-10
 
 - [Added] Document how to use --max-old-space-size