+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2020-04-03
+
+- [Changed] Updated dependencies in package.json
+- [Changed] TestBed.get -> TestBed.inject (Angular 9 deprecation)
 - [Changed] Only import HttpClientModule in appModule
 
 ## [1.1.0] - 2020-02-10